+++ character/models/races.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from ..constants.races import LanguageName, LanguageType, SenseName
+from ..constants.races import LanguageName, LanguageType
 
 
 class Language(models.Model):
@@ -9,17 +9,6 @@ class Language(models.Model):
     script = models.CharField(
         max_length=11, choices=LanguageName.choices, blank=True, null=True
     )
-
-    def __str__(self):
-        return str(self.name)
-
-
-class Sense(models.Model):
-    name = models.CharField(max_length=30, choices=SenseName.choices, unique=True)
-    description = models.TextField(max_length=50)
-
-    class Meta:
-        ordering = ["name"]
 
     def __str__(self):
         return str(self.name)